HRCPP-151 Do not use STL classes in public API * std::* in classes passed over the library boundary replaced with portable::* ** internally the client still uses STL, public API also takes STL but rewraps it * more careful reasoning about allocation/deallocation ** object allocated in one module has to be released in the same module ** that's why explicit destructor functions needs to be passed * most of the changes are public API-wise compatible * hrbytes hidden as internal class * Unit.cpp, L3Test.cpp and HashTest.cpp moved into the library, exporting the functions * marshalling calls moved from RemoteCache to RemoteCacheImpl (to call all marshalling/unmarshalling from one place only) * a bit of headers cleanup * explicit casts for integer type conversions (to remove warnings on Windows) - but no changes in API

+++ src/hotrod/sys/BlockingQueue.h
@@ -71,7 +71,7 @@ public:
         return old_size != queue.size();
     }
 
-    int size() {
+    size_t size() {
         ScopedLock<Mutex> l(lock);
 
         return queue.size();
